fix: refact volume manage routes
