+++ src/pages/clusters/routes/index.js
@@ -57,7 +57,6 @@ import NetworkPolicies from '../containers/Network/Policies'
 import IPPools from '../containers/Network/IPPools'
 import LogCollections from '../containers/LogCollections'
 import CustomMonitoring from '../containers/CustomMonitoring'
-import PV from '../containers/Storage/PV'
 import detail from './detail'
 
 const PATH = '/clusters/:cluster'
@@ -181,12 +180,6 @@ export default [
           {
             path: `${PATH}/volumes`,
             component: Volumes,
-            exact: true,
-          },
-          {
-            path: `${PATH}/pv`,
-            component: PV,
-            exact: true,
           },
           {
             path: `${PATH}/volume-snapshots`,
